Add left sidebar for board selection and split Board.vue
