+++ vue/src/components/EditorSidebar.vue
@@ -20,7 +20,7 @@
     <div v-else-if="editorId === 3">
       <PollEditor @create-poll="createPoll"/>
     </div>
-    <br><br><br> <!-- For scrollbar -->
+    <br><br><br><br> <!-- For scrollbar -->
   </div>
 </template>
 
@@ -33,6 +33,7 @@ import NoteEditor from '@/components/editors/NoteEditor.vue'
 import PollEditor from '@/components/editors/PollEditor.vue'
 
 export default {
+  name: 'BoardSidebar',
   components: {
     EditorHeader,
     NoteEditor,
@@ -85,11 +86,21 @@ export default {
         })
     },
     createNote: async function (titleContent, jsonContent) {
-      const jsonBody = JSON.stringify({
-        title: titleContent,
-        typeOfPost: 'application/note',
-        content: jsonContent
-      })
+      var jsonBody
+      if (this.dueDate == null) {
+        jsonBody = JSON.stringify({
+          title: titleContent,
+          typeOfPost: 'application/note',
+          content: jsonContent
+        })
+      } else {
+        jsonBody = JSON.stringify({
+          title: titleContent,
+          typeOfPost: 'application/note',
+          dueDate: this.dueDate,
+          content: jsonContent
+        })
+      }
 
       // Post request to api
       this.refreshToken()
@@ -112,11 +123,22 @@ export default {
       const datapart = dataURISplit[0] // E.g. data:image/png;base64
       const base64Data = dataURISplit[1] // BORw0KGgoAAAANSUhEUgAAB...
       const dataType = datapart.split(':')[1].split(';')[0] // Extract MIME type -> image/png
-      const jsonBody = JSON.stringify({
-        title: titleContent,
-        typeOfPost: dataType,
-        content: base64Data
-      })
+
+      var jsonBody
+      if (this.dueDate == null) {
+        jsonBody = JSON.stringify({
+          title: titleContent,
+          typeOfPost: dataType,
+          content: base64Data
+        })
+      } else {
+        jsonBody = JSON.stringify({
+          title: titleContent,
+          typeOfPost: dataType,
+          dueDate: this.dueDate,
+          content: base64Data
+        })
+      }
 
       // Post request to api
       this.refreshToken()
@@ -139,11 +161,22 @@ export default {
       const datapart = dataURISplit[0] // e.g. data:application/pdf;base64
       const base64Data = dataURISplit[1] // ZGgoAAAANSUhEUgAASs54B...
       const dataType = datapart.split(':')[1].split(';')[0] // Extract MIME type -> application/pdf
-      const jsonBody = JSON.stringify({
-        title: titleContent,
-        typeOfPost: dataType,
-        content: base64Data
-      })
+
+      var jsonBody
+      if (this.dueDate == null) {
+        jsonBody = JSON.stringify({
+          title: titleContent,
+          typeOfPost: dataType,
+          content: base64Data
+        })
+      } else {
+        jsonBody = JSON.stringify({
+          title: titleContent,
+          typeOfPost: dataType,
+          dueDate: this.dueDate,
+          content: base64Data
+        })
+      }
 
       // Post request to api
       this.refreshToken()
@@ -162,11 +195,21 @@ export default {
       this.$emit('add-note')
     },
     createPoll: async function (titleContent, jsonContent, answerIndexes) {
-      const jsonBodyNote = JSON.stringify({
-        title: titleContent,
-        typeOfPost: 'application/poll',
-        content: jsonContent
-      })
+      var jsonBodyNote
+      if (this.dueDate == null) {
+        jsonBodyNote = JSON.stringify({
+          title: titleContent,
+          typeOfPost: 'application/poll',
+          content: jsonContent
+        })
+      } else {
+        jsonBodyNote = JSON.stringify({
+          title: titleContent,
+          typeOfPost: 'application/poll',
+          dueDate: this.dueDate,
+          content: jsonContent
+        })
+      }
 
       // Post request to api
       this.refreshToken()
